Added drafts of classes in scratch.py.
Components are the basic parent class for classes. Contains metadata
attributes name and description and a "show" class to print out the
components in easy formatting.

Beacons are a child class from Components. Beacons are a logical class
that corresponds to a physical RFID tag in the user's environment. This
class adds location and code attributes and a "scan" method to wait for
the user to scan RFID tags. Scan compares the input code and the code
stored in the beacon object and loops if they do not match.

Actions are also a child of the Component class. The actions class
reflects a discreet action that the user will take: going to a beacon,
completing single action.

Tasks are also a child of the Component Class. Tasks are groups of
common Actions that the user will build and execute to complete a single
task. Tasks have Component metadata and a list of Action objects.