Implemented single routine logic in main.py
Main.py prompts the user to create a single Routine using the terminal
to input information. For the Routine, the program prompts the user to
create 3 Tasks, each with 3 Steps, each with one Waypoint.

Next the program prompts the user to follow the Routine by executing
each Step of each Task in the routine. For now, Waypoints are inputted
by getting terminal inputs for a code string (all are default set to
zeros) and checking it against the code saved in the Waypoint object.

The script main.py contains wrapper functions that get user inputs and
return objects or tuples with information for object creation. The main
function loops through the task_limit to create tasks Tasks and Steps.